bot/requirement.txt
discord.py==1.7.3
python-dotenv==0.19.2
pytest==7.0.1